Bruce Evans:
I still don't understand the overflow possibilities mentioned next to
the XXX.  errtmp was added by my patch to 0.0.  Bill deleted the
volatile but it is necessary to stop gcc -O from optimizing away the
assignments to errtmp.  Of course volatile is not portable to old
compilers.  No doubt that's why it got deleted.  My minix version of the
library has volatile errtmp's in several other files but my 386BSD
version only has them for pow.c (they came with this fix).  The point of
the errtmp expressions was to set various IEEE flags but we are a long
way from supporting this properly.  First we need to support ANSI
behaviour.