>From marc@escargot.rain.com Fri Sep 10 00:04:44 1993
BTW, one thing I noticed is that the syslog() calls in slattach add
a newline to the end of the string. Looks like syslog() automatically
already provides a newline, so extra blank lines appear on the console
display causing other useful important information to scroll off
sooner. If you want me to send you a diff when I'm done, I can do that,
otherwise the newlines in the various syslog() statements could be
removed... Comments?

Removed.