Patch from Andy Heffernan <ahh@netcom.com> Description from Sean Eric Fagan:
This has never bitten me, but someone (whose name I forget, but I do
thank) pointed out it was wrong.  He was right, so here is a fix.
(Basicly, the kernel's address of the temporary page gets set by
one of the vm* routines, so I didn't bother initializing it, but it
does get checked by the vm* routine before it sets it, so it should be
initialized.  Zero seems a reasonable value.)  (Oh, and if whomever-