>From Jordan K. Hubbard >From: jtk@atria.com (John T. Kohl)
in rcmd:
It calls select() with a hardcoded "number of file descriptors" argument
of 32, rather than computing it based on the sockets about which it
cares.

- change this to use FD_SETSIZE -- quick and dirty - that's the largest we can
  go in a FD_SET

I suggest we look at fixing these arguments to select() better in 2.0 --
there are lots of suspect looking calls to select in the 1.x tree.
I just don't have time to do a proper fix now.

Also: add some typecasts to shut the compiler up:
	-- setgroups(2) wants an (int *) not a (gid_t *) as it's second
	   argument.